Refactor ChainFactory and create VectorDBManager
